updated models and routes
