+++ server/controllers/user_controller.go
@@ -1,0 +1,1 @@
+package controllers
