+++ server/routes/user_routes.go
@@ -1,0 +1,13 @@
+package routes
+
+import (
+	"github.com/gin-gonic/gin"
+	"remind/server/controllers"
+)
+
+func UserRoutes(r *gin.Engine) {
+	userGroup := r.Group("/user")
+	{
+		userGroup.GET("")
+	}
+}
