+++ ui/src/app/types/reminder.t.ts
@@ -1,0 +1,9 @@
+
+
+export type reminder = {
+  title: string
+  description: string
+
+}
+
+
